requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
supabase>=2.4.0
snowflake-connector-python>=3.10.0
pandas>=2.2.0
numpy>=1.26.0
orjson>=3.9.10
streamlit>=1.40.0
plotly>=5.24.0
